requirements.txt
django==3.2.16
djangorestframework==3.14.0
python-decouple==3.6
djangorestframework-simplejwt==5.2.2
django-filter==22.1
pillow==9.3.0
psycopg2-binary==2.9.5